requirements.txt
python-dotenv==1.0.1
colorama==0.4.6